Refactor provider credentials store
