+++ packages/forklift-console-plugin/src/modules/Providers/views/details/components/CredentialsSection/components/state/index.ts
@@ -1,0 +1,3 @@
+// @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './reducer';
+// @endindex
